// src/app/shared/models/qcp-list.ts
export class QcpList {
  id: string;
  cqmId: string;
  projectId: string;
  name: string;
  riserTag: string;
  riserModel: string;
  subAssembly: string;
  componentName: string;
  componentDescription: string;
  recordName: string;
  recordDescription: string;
  acceptanceCriteria: string;
  referenceProcedure: string;
  manualReport: boolean;
  clientName: string;
  rigName: string;
  projectName: string;
  user: string;
}